requirements.txt
Django==3.2.6
django-crispy-forms==1.12.0
django-taggit==1.5.1
dj_database_url==0.5.0
colorama==0.4.3
keyring==18.0.1
psycopg2_binary==2.9.1
backports.statistics==0.1.0
bpython==0.21
contextvars==2.4
cx_Oracle==8.2.1
Cython==0.29.24
djangorestframework==3.12.4
docutils==0.17.1
ipaddr==2.2.0
ipdb==0.13.9
ipython==7.27.0
mypy_extensions==0.4.3
numpy==1.21.2
ordereddict==1.1
Pillow==8.3.2
psycopg2==2.9.1
pylibmc==1.6.1
pymemcache==3.5.0
python_memcached==1.59
pywatchman==1.4.1
PyYAML==5.4.1
SecretStorage==2.3.1
tblib==1.7.0
testsuite==0.1.2
typing_extensions==3.10.0.2
Unidecode==1.3.1
wincertstore==0.2
gunicorn
whitenoise